Extracts `FeedViewController` test helpers to a separate file
